Add test caching functionality
Modifies Config to add a new TestConfig containing a `cache_dir`, a
directory to store test results.

When testing chapters, the modified time of the source file is compared
to the modified time of a corresponding file in the cache directory. If
the source is newer, then the tests are run.

After running a successful test, the cache file is created.